solver/requirements.txt
Flask==1.1.2
py-sudoku==1.0.3